html-app-0.0.1/requirements.txt
Flask==1.1.1
requests==2.22.0
PyJWT==1.5.3